+++ Documentation/Changelog.md
@@ -31,6 +31,7 @@ Release date: not released yet
   - Added polyfills for `Split` and `SplitAny` methods of `System.ReadOnlySpan<char>` type
   - Added polyfill for `ReadSpan` method of `System.Runtime.InteropServices.SafeBuffer` type
   - Added polyfill for `System.Runtime.CompilerServices.CallerArgumentExpressionAttribute` type
+  - Added polyfills for `System.Threading.Tasks.Task.WaitAsync` and `System.Threading.Tasks.Task<TResult>.WaitAsync` methods
   - Added polyfill for a LINQ-style pipe operator
   - Removed `System.HashCode` polyfill implementation in favor of `Microsoft.Bcl.HashCode` package
 - .NET Framework 4.6 support is retired. The minimal supported version of .NET Framework is 4.6.1